+++ lib/mongo/unordered_bulk.ex
@@ -6,7 +6,17 @@ defmodule Mongo.UnorderedBulk do
   """
 
   alias Mongo.UnorderedBulk
-  import Mongo.BulkUtils
+  alias Mongo.BulkWrite
+
+  import Mongo.BulkOps
+  import Mongo.Utils
+
+  @type t :: %__MODULE__{
+               coll: String.t,
+               inserts: [BulkOps.bulk_op],
+               updates: [BulkOps.bulk_op],
+               deletes: [BulkOps.bulk_op]
+             }
 
   defstruct coll: nil, inserts: [], updates: [], deletes: []
 
@@ -37,43 +47,27 @@ defmodule Mongo.UnorderedBulk do
   end
 
   def replace_one(%UnorderedBulk{} = bulk, filter, replacement, opts \\ []) do
-    _ = modifier_docs(replacement, :replace)
     get_replace_one(filter, replacement, opts) |> push(bulk)
   end
 
   def update_one(%UnorderedBulk{} = bulk, filter, update, opts \\ []) do
-    _ = modifier_docs(update, :update)
     get_update_one(filter, update, opts) |> push(bulk)
   end
 
-  def update_many(%UnorderedBulk{updates: rest} = bulk, filter, update, opts \\ []) do
-    _ = modifier_docs(update, :update)
+  def update_many(%UnorderedBulk{} = bulk, filter, update, opts \\ []) do
     get_update_many(filter, update, opts) |> push(bulk)
   end
 
-  defp modifier_docs([{key, _}|_], type), do: key |> key_to_string |> modifier_key(type)
-  defp modifier_docs(map, _type) when is_map(map) and map_size(map) == 0,  do: :ok
-  defp modifier_docs(map, type) when is_map(map), do: Enum.at(map, 0) |> elem(0) |> key_to_string |> modifier_key(type)
-  defp modifier_docs(list, type) when is_list(list),  do: Enum.map(list, &modifier_docs(&1, type))
-
-  defp modifier_key(<<?$, _::binary>> = other, :replace),  do: raise(ArgumentError, "replace does not allow atomic modifiers, got: #{other}")
-  defp modifier_key(<<?$, _::binary>>, :update),  do: :ok
-  defp modifier_key(<<_, _::binary>> = other, :update),  do: raise(ArgumentError, "update only allows atomic modifiers, got: #{other}")
-  defp modifier_key(_, _),  do: :ok
-
-  defp key_to_string(key) when is_atom(key), do: Atom.to_string(key)
-  defp key_to_string(key) when is_binary(key), do: key
-
-  def stream(enum, top, coll, limit \\ 1000, opts \\ []) when limit > 1 do
+  def write(enum, top, coll, limit \\ 1000, opts \\ []) when limit > 1 do
     Stream.chunk_while(enum,
       {new(coll), limit - 1},
       fn
-        op, {bulk, 0} -> {:cont, Mongo.BulkWrite.bulk_write(top, push(op, bulk), opts), {new(coll), limit - 1}}
+        op, {bulk, 0} -> {:cont, BulkWrite.write(top, push(op, bulk), opts), {new(coll), limit - 1}}
         op, {bulk, l} -> {:cont, {push(op, bulk), l - 1}}
       end,
       fn
         {bulk, 0} -> {:cont, bulk}
-        {bulk, _} -> {:cont, Mongo.BulkWrite.bulk_write(top, bulk, opts), {new(coll), limit - 1}}
+        {bulk, _} -> {:cont, BulkWrite.write(top, bulk, opts), {new(coll), limit - 1}}
     end)
     # todo reduce to one
   end
@@ -92,7 +86,7 @@ defmodule Mongo.UnorderedBulk do
     |> delete_one(%{kind: "dog"})
     |> delete_one(%{kind: "dog"})
 
-    result = Mongo.BulkWrite.bulk_write(top, bulk, w: 1)
+    result = BulkWrite.write(top, bulk, w: 1)
 
     IO.puts inspect result
   end
@@ -109,7 +103,7 @@ defmodule Mongo.UnorderedBulk do
            |> replace_one(%{name: "Tom"}, %{name: "Tom", kind: "dog"})
            |> delete_many(%{kind: "dog"})
 
-    result = Mongo.BulkWrite.bulk_write(top, bulk, w: 1)
+    result = BulkWrite.write(top, bulk, w: 1)
 
     IO.puts inspect result
   end
@@ -125,7 +119,7 @@ defmodule Mongo.UnorderedBulk do
            |> update_many(%{name: %{"$exists": true}}, %{"$set": %{kind: "dog"}})
            |> delete_many(%{kind: "dog"})
 
-    result = Mongo.BulkWrite.bulk_write(top, bulk, w: 1)
+    result = BulkWrite.write(top, bulk, w: 1)
 
     IO.puts inspect result
   end