+++ CHANGELOG.md
@@ -1,6 +1,12 @@
 ## v0.5.4
 
 * Enhancements
+   * Added support for bulk writes (ordered/unordered)
+   * Added support for op_msg with payload type 1 
+
+## v0.5.4
+
+* Enhancements
    * The driver provides now client metadata 
    * Added support for connecting via UNIX sockets (`:socket` and `:socket_dir`)